fix(UT-383): share the suite Redis with collected Django tests
Imported TaskTestCase runs only saw the testcontainer when package tests
started it first. Session Redis now lives at the project root so a single
Django case starts the same Redis; drop the unused checkout override and
keep download failures named with version and cache path.
